extract into fn
