+++ src/internal/images/pull.go
@@ -1,9 +1,13 @@
 package images
 
 import (
+	"crypto/sha256"
+	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
+	"os"
+	"path"
 	"path/filepath"
 	"strings"
 
@@ -106,4 +110,79 @@ func PullAll(buildImageList []string, imageTarballPath string) map[name.Tag]v1.I
 	}
 
 	return tagToImage
+}
+
+func FormatCraneOCILayout(ociPath string) error {
+	type IndexJSON struct {
+		SchemaVersion int `json:"schemaVersion"`
+		Manifests     []struct {
+			MediaType string `json:"mediaType"`
+			Size      int    `json:"size"`
+			Digest    string `json:"digest"`
+		} `json:"manifests"`
+	}
+
+	indexJson, err := os.Open(path.Join(ociPath, "index.json"))
+	if err != nil {
+		message.Errorf(err, "Unable to open %s/index.json", ociPath)
+		return err
+	}
+	var index IndexJSON
+	byteValue, _ := io.ReadAll(indexJson)
+	json.Unmarshal(byteValue, &index)
+
+	digest := strings.TrimPrefix(index.Manifests[0].Digest, "sha256:")
+	b, err := os.ReadFile(path.Join(ociPath, "blobs", "sha256", digest))
+	if err != nil {
+		message.Errorf(err, "Unable to open %s/blobs/sha256/%s", ociPath, digest)
+		return err
+	}
+	manifest := string(b)
+	// replace all docker media types w/ oci media types
+	manifest = strings.ReplaceAll(manifest, "application/vnd.docker.distribution.manifest.v2+json", "application/vnd.oci.image.manifest.v1+json")
+	manifest = strings.ReplaceAll(manifest, "application/vnd.docker.image.rootfs.diff.tar.gzip", "application/vnd.oci.image.layer.v1.tar+gzip")
+
+	h := sha256.New()
+	h.Write([]byte(manifest))
+	bs := h.Sum(nil)
+
+	// Write the manifest to the blobs directory w/ the sha256 hash as the filename
+	manifestPath := path.Join(ociPath, "blobs", "sha256", fmt.Sprintf("%x", bs))
+	manifestFile, err := os.Create(manifestPath)
+	if err != nil {
+		message.Errorf(err, "Unable to create %s/blobs/sha256/%x", ociPath, bs)
+		return err
+	}
+	defer manifestFile.Close()
+	_, err = manifestFile.WriteString(manifest)
+	if err != nil {
+		message.Errorf(err, "Unable to write to %s/blobs/sha256/%x", ociPath, bs)
+		return err
+	}
+
+	// Update the index.json to point to the new manifest
+	index.SchemaVersion = 2
+	index.Manifests[0].Digest = fmt.Sprintf("sha256:%x", bs)
+	index.Manifests[0].Size = len(manifest)
+	index.Manifests[0].MediaType = "application/vnd.oci.image.manifest.v1+json"
+	indexJson.Close()
+	_ = os.Remove(path.Join(ociPath, "index.json"))
+	indexJson, err = os.Create(path.Join(ociPath, "index.json"))
+	if err != nil {
+		message.Errorf(err, "Unable to create %s/index.json", ociPath)
+		return err
+	}
+	indexJsonBytes, err := json.Marshal(index)
+	if err != nil {
+		message.Errorf(err, "Unable to marshal %s/index.json", ociPath)
+		return err
+	}
+	_, err = indexJson.Write(indexJsonBytes)
+	if err != nil {
+		message.Errorf(err, "Unable to write to %s/index.json", ociPath)
+		return err
+	}
+	indexJson.Close()
+
+	return nil
 }
